S104 split: polaris + instance_orchestrator
